rendering user profile and images finally
