+++ client/src/pages/SwipePage.js
@@ -8,6 +8,7 @@ import'swiper/swiper-bundle.css';
 import Profile from '../components/Profile/Profile';
 import Loading from '../components/Loading';
 import '../css/SwipePage.css';
+import auth from '../services/auth.js';
 
 SwiperCore.use([Navigation]);
 
@@ -17,7 +18,7 @@ class SwipePage extends React.Component {
   state = {
     profiles: [],
     loading: true,
-    userID: null,
+    id: null,
   }
 
   componentDidMount() {
@@ -25,7 +26,7 @@ class SwipePage extends React.Component {
     .then(res => res.json())
     .then(profiles => {
       this.setState({
-        userID: this.props.match.params,
+        id: auth.currentUser.id,
         loading: false,
         profiles: profiles.map((p, ii) => 
           <Profile {...p} key={ii} />
@@ -40,22 +41,20 @@ class SwipePage extends React.Component {
       return <Loading />
     
     let slides = []; //we needed a new array to be modified, because we cannot modify the state itself without setState.  
-    
+    console.log(this.state.profiles[1].res);
+
     for (let i=0; i<this.state.profiles.length; i+=1){
-      // if(this.state.profiles[i].Profile.id != {this.state.userID})
-      slides.push(
-        <SwiperSlide key={`slide-${i}`} tag="ul">
-          <div class='rectangle'>
-            {this.state.profiles[i]}
-          </div>  
-        </SwiperSlide>
-      );
+      if(this.state.profiles[i].id !== this.state.id){
+
+        slides.push(
+          <SwiperSlide key={`slide-${i}`} tag="ul">
+            <div class='rectangle'>
+              {this.state.profiles[i]}
+            </div>  
+          </SwiperSlide>
+        );
+      }
 //after everyone has been swiped, transfer them to another array of already been swiped.
-
-//Notes for possibly cyclic loop
-//look at Swiper library in React and they prob have a way of detecting if you swiped left or right
-// look for default behavior to bring the next slides
-
     }
    
     return (
