+++ package.json
@@ -17,6 +17,7 @@
     "clipboard": "^2.0.8",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
+    "framer": "^1.2.30",
     "morgan": "^1.10.0",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
